Refactor waitlist page into separate component
We'll be reusing that page with only a couple of strings modified,
so it felt OK to abstract that out.
